Link against major.minor.patch version of libmysqlclient to avoid runtime conflict
Summary: Someone shipped a libmysqlclient.so.21 in the system but that was of older version that didn't have mysql_result_metadata (actually it did - but locally and not exported). As a result libmysql_test_api is failing to locate mysql_result_metadata at runtime. The fix is to always link against major.minor.patch version that we are building to avoid conflicts.

Test Plan:
Before - you'll see this when you do a build:

```
./runtime_output_directory/libmysql_api_test: relocation error: ../runtime_output_directory/libmysql_api_test: symbol mysql_result_metadata, version libmysqlclient_21.0 not defined in file libmysqlclient.so.21 with link time reference
```

Now: It builds

Reviewers: zhichengzhu, #mysql_eng

Reviewed By: zhichengzhu

Subscribers: zhichengzhu, ritwikyadav, vinaybhat, webscalesql-eng@fb.com

Differential Revision: https://phabricator.intern.facebook.com/D17507965

Signature: 17507965:1569012155:d3f4f20fdfbadf7ec9a89a1e4149a621196602d6